requirements.txt
Django==1.6
South==0.8.3
argparse==1.2.1
dj-database-url==0.2.2
django-chartit==0.1
oauthlib==0.6.0
psycopg2==2.5.1
requests==2.0.0
requests-oauthlib==0.3.2
simplejson==3.3.1
twython==3.1.0
wsgiref==0.1.2
